Widen SAP pair-count arithmetic to avoid signed overflow.
The three pair buffers are sized as n*(n-1)/2 in int. The product
overflows once n reaches 46342, although the halved result fits int for
every n that mj_SAP accepts. The wrapped count is negative and
mjSTACKALLOC fails, so a 1D/2D flex with 46342 to 65535 elements and
selfcollide="sap" cannot be simulated, even though its buffer is capped
at 1000000 pairs and would fit.

Computing the product in size_t makes it exact. The compiler limits
bodies plus flexes to 65533 and mj_SAP rejects n >= 0x10000, so the
result fits int when narrowed back.

The regression test now builds 46342 elements, which covers this
overflow too. It takes 0.8 s instead of 0.4 s.

+++ src/engine/engine_collision_driver.c
@@ -618,7 +618,7 @@ void mj_collision(const mjModel* m, mjData* d) {
 
   // broadphase collision detector
   TM_START;
-  int nmaxpairs = (nbodyflex*(nbodyflex - 1))/2;
+  int nmaxpairs = (size_t)nbodyflex*(nbodyflex - 1)/2;
   uint32_t* broadphasepair = mjSTACKALLOC(d, nmaxpairs, uint32_t);
   int nbfpair = mj_broadphase(m, d, broadphasepair, nmaxpairs);
   unsigned int last_signature = -1;
@@ -1679,7 +1679,7 @@ int mj_broadphase(const mjModel* m, mjData* d, uint32_t* bfpair, int maxpair) {
     }
 
     // call SAP
-    int maxsappair = ncollide*(ncollide-1)/2;
+    int maxsappair = (size_t)ncollide*(ncollide-1)/2;
     uint32_t* sappair = mjSTACKALLOC(d, maxsappair, uint32_t);
     int nsappair = mj_SAP(d, aamm, ncollide, 0, sappair, maxsappair);
     if (nsappair < 0) {
@@ -2349,7 +2349,7 @@ void mj_collideFlexSAP(const mjModel* m, mjData* d, int f) {
   int axis = (bvh[3] > bvh[4] && bvh[3] > bvh[5]) ? 0 : (bvh[4] > bvh[5] ? 1 : 2);
 
   // call SAP; hard limit on number of pairs to avoid out-of-memory
-  int maxsappair = mjMIN(nactive*(nactive-1)/2, 1000000);
+  int maxsappair = mjMIN((size_t)nactive*(nactive-1)/2, 1000000);
   uint32_t* sappair = mjSTACKALLOC(d, maxsappair, uint32_t);
   int nsappair = mj_SAP(d, aamm, nactive, axis, sappair, maxsappair);
   if (nsappair < 0) {
